refactor: disable multiplexer per build flag
