+++ cmd/celestia-appd/cmd/root.go
@@ -5,22 +5,14 @@ import (
 	"os"
 
 	"cosmossdk.io/log"
-	confixcmd "cosmossdk.io/tools/confix/cmd"
-	"github.com/01builders/nova"
-	"github.com/cometbft/cometbft/cmd/cometbft/commands"
-	tmcli "github.com/cometbft/cometbft/libs/cli"
 	dbm "github.com/cosmos/cosmos-db"
 	"github.com/cosmos/cosmos-sdk/client"
 	clientconfig "github.com/cosmos/cosmos-sdk/client/config"
-	"github.com/cosmos/cosmos-sdk/client/debug"
 	"github.com/cosmos/cosmos-sdk/client/flags"
-	"github.com/cosmos/cosmos-sdk/client/keys"
-	"github.com/cosmos/cosmos-sdk/client/snapshot"
 	"github.com/cosmos/cosmos-sdk/server"
 	serverconfig "github.com/cosmos/cosmos-sdk/server/config"
 	simtestutil "github.com/cosmos/cosmos-sdk/testutil/sims"
 	"github.com/cosmos/cosmos-sdk/x/auth/types"
-	genutilcli "github.com/cosmos/cosmos-sdk/x/genutil/client/cli"
 	kitlog "github.com/go-kit/log"
 	"github.com/spf13/cobra"
 
@@ -108,59 +100,6 @@ func NewRootCmd() *cobra.Command {
 	}
 
 	return rootCommand
-}
-
-// initRootCommand performs a bunch of side-effects on the root command.
-func initRootCommand(rootCommand *cobra.Command, capp *app.App) {
-	versions := Versions()
-
-	debugCmd := debug.Cmd()
-	debugCmd.AddCommand(
-		NewInPlaceTestnetCmd(),
-		AppGenesisToCometGenesisConverterCmd(),
-	)
-	passthroughCmd, _ := nova.NewPassthroughCmd(versions)
-	// TODO: handle the error here. (currently breaking ledger tests as they do a cli exec and the expected binary bytes are not there)
-
-	rootCommand.AddCommand(
-		genutilcli.InitCmd(capp.BasicManager, app.DefaultNodeHome),
-		genutilcli.Commands(capp.GetTxConfig(), capp.BasicManager, app.DefaultNodeHome),
-		tmcli.NewCompletionCmd(rootCommand, true),
-		debugCmd,
-		confixcmd.ConfigCommand(),
-		commands.CompactGoLevelDBCmd,
-		addrbookCommand(),
-		downloadGenesisCommand(),
-		addrConversionCmd(),
-		server.StatusCommand(),
-		queryCommand(capp.BasicManager),
-		txCommand(capp.BasicManager),
-		keys.Commands(),
-		snapshot.Cmd(NewAppServer),
-		passthroughCmd,
-	)
-
-	// Add the following commands to the rootCommand: start, tendermint, export, version, and rollback.
-	server.AddCommandsWithStartCmdOptions(rootCommand, app.DefaultNodeHome, NewAppServer, appExporter, server.StartCmdOptions{
-		AddFlags:            addStartFlags,
-		StartCommandHandler: nova.New(versions), // multiplexer
-	})
-
-	// find start command
-	startCmd, _, err := rootCommand.Find([]string{"start"})
-	if err != nil {
-		panic(fmt.Errorf("failed to find start command: %w", err))
-	}
-	startCmdRunE := startCmd.RunE
-
-	// Add the BBR check to the start command
-	startCmd.RunE = func(cmd *cobra.Command, args []string) error {
-		if err := checkBBR(cmd); err != nil {
-			return err
-		}
-
-		return startCmdRunE(cmd, args)
-	}
 }
 
 // addStartFlags adds flags to the start command.
